+++ docs/changelog.md
@@ -28,6 +28,7 @@ pageflow_next_text: Contributing
   - `EnsureAllOf` now implements `ListAssurance` interface
   - `EnsureAnyOneOf` now implements `ListAssurance` interface
   - `RequireAllOf` now implements `ListRequirement` interface
+  - `RequireAnyOneOf` now implements `ListRequirement` interface
 
 ## v1.2016062801